requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
asttokens==3.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.40.27
botocore==1.40.27
cachetools==5.5.2
cellpose==2.2.3
CellProfiler==4.2.8
cellprofiler-core==4.2.8
centrosome==1.2.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.1.1
cycler==0.12.1
databricks-sdk==0.65.0
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docutils==0.15.2
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
fastremap==1.17.4
filelock==3.19.1
fill_voids==2.1.1
fire==0.7.1
Flask==3.1.2
fonttools==4.59.2
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.31.0
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.6.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.2
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
llvmlite==0.43.0
mahotas==1.4.18
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mlflow==3.1.4
mlflow-skinny==3.1.4
monai==1.5.0
mpmath==1.3.0
mysqlclient @ file:///home/conda/feedstock_root/build_artifacts/mysqlclient_1609783192600/work
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.107.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prokaryote==2.4.4
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pynvml==13.0.1
pyparsing==3.2.3
python-bioformats==4.1.0
python-dateutil==2.9.0.post0
python-javabridge==4.0.4
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==22.3.0
regex==2025.9.1
requests==2.32.5
roifile==2024.9.15
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.18.3
scikit-learn==1.6.1
scipy==1.10.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tokenizers==0.22.0
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
